yolact_cuda_free/requirement.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow==1.14
keras==2.3.1
opencv-python
h5py
imgaug
scikit-learn
IPython[all]
pytorch
boto3
botocore
flask